--- a/loser/loser.go
+++ b/loser/loser.go
@@ -61,8 +61,11 @@ func (t *Tree[E]) Next() bool {
 	if t.nodes[t.nodes[0].index].index == -1 { // already exhausted
 		return false
 	}
-	t.moveNext(t.nodes[0].index)
-	t.replayGames(t.nodes[0].index)
+	if t.moveNext(t.nodes[0].index) {
+		t.replayGames(t.nodes[0].index)
+	} else {
+		t.sequenceEnded(t.nodes[0].index)
+	}
 	return t.nodes[t.nodes[0].index].index != -1
 }
 
@@ -84,12 +87,14 @@ func (t *Tree[E]) initialize() {
 	t.nodes[0].value = t.nodes[winners[1]].value
 }
 
-// Starting at pos, re-consider all values up to the root.
+// Starting at pos, which is a winner, re-consider all values up to the root.
 func (t *Tree[E]) replayGames(pos int) {
 	// At the start, pos is a leaf node, and is the winner at that level.
 	n := parent(pos)
 	for n != 0 {
-		if t.nodes[n].value <= t.nodes[pos].value {
+		// If n.value < pos.value then pos loses.
+		// If they are equal, pos wins because n could be a sequence that ended, with value maxval.
+		if t.nodes[n].value < t.nodes[pos].value {
 			loser := pos
 			// Record pos as the loser here, and the old loser is the new winner.
 			pos = t.nodes[n].index
@@ -103,8 +108,29 @@ func (t *Tree[E]) replayGames(pos int) {
 	t.nodes[0].value = t.nodes[pos].value
 }
 
+func (t *Tree[E]) sequenceEnded(pos int) {
+	// Find the first active sequence which used to lose to it.
+	n := parent(pos)
+	for n != 0 && t.nodes[t.nodes[n].index].index == -1 {
+		n = parent(n)
+	}
+	if n == 0 {
+		// There are no active sequences left
+		t.nodes[0].index = pos
+		t.nodes[0].value = t.maxVal
+		return
+	}
+
+	// Record pos as the loser here, and the old loser is the new winner.
+	loser := pos
+	winner := t.nodes[n].index
+	t.nodes[n].index = loser
+	t.nodes[n].value = t.nodes[loser].value
+	t.replayGames(winner)
+}
+
 func (t *Tree[E]) playGame(a, b int) (loser, winner int) {
-	if t.nodes[a].value <= t.nodes[b].value {
+	if t.nodes[a].value < t.nodes[b].value {
 		return b, a
 	}
 	return a, b