--- a/loser/loser_test.go
+++ b/loser/loser_test.go
@@ -30,14 +30,6 @@ func (it *List) Next() bool {
 	return false
 }
 
-func (it *List) Seek(val uint64) bool {
-	for it.cur < val && len(it.list) > 0 {
-		it.cur = it.list[0]
-		it.list = it.list[1:]
-	}
-	return len(it.list) > 0
-}
-
 func checkIterablesEqual[E any, S1 loser.Sequence, S2 loser.Sequence](t *testing.T, a S1, b S2, at1 func(S1) E, at2 func(S2) E, less func(E, E) bool) {
 	t.Helper()
 	count := 0