--- a/ring/model.go
+++ b/ring/model.go
@@ -1,8 +1,9 @@
 package ring
 
 import (
-	"container/heap"
 	"fmt"
+	"github.com/grafana/dskit/loser"
+	"math"
 	"sort"
 	"sync"
 	"time"
@@ -590,32 +591,27 @@ func GetOrCreateRingDesc(d interface{}) *Desc {
 	return d.(*Desc)
 }
 
-// TokensHeap is an heap data structure used to merge multiple lists
-// of sorted tokens into a single one.
-type TokensHeap [][]uint32
-
-func (h TokensHeap) Len() int {
-	return len(h)
+type tokenList struct {
+	list []uint32
+	cur  uint32
 }
 
-func (h TokensHeap) Swap(i, j int) {
-	h[i], h[j] = h[j], h[i]
+func newTokenList(list ...uint32) *tokenList {
+	return &tokenList{list: list}
 }
 
-func (h TokensHeap) Less(i, j int) bool {
-	return h[i][0] < h[j][0]
+func (it *tokenList) At() uint32 {
+	return it.cur
 }
 
-func (h *TokensHeap) Push(x interface{}) {
-	*h = append(*h, x.([]uint32))
-}
-
-func (h *TokensHeap) Pop() interface{} {
-	old := *h
-	n := len(old)
-	x := old[n-1]
-	*h = old[0 : n-1]
-	return x
+func (it *tokenList) Next() bool {
+	if len(it.list) > 0 {
+		it.cur = it.list[0]
+		it.list = it.list[1:]
+		return true
+	}
+	it.cur = 0
+	return false
 }
 
 // MergeTokens takes in input multiple lists of tokens and returns a single list
@@ -623,35 +619,26 @@ func (h *TokensHeap) Pop() interface{} {
 // to have tokens already sorted.
 func MergeTokens(instances [][]uint32) []uint32 {
 	numTokens := 0
+	lists := make([]*tokenList, len(instances))
 
-	// Build the heap.
-	h := make(TokensHeap, 0, len(instances))
-	for _, tokens := range instances {
+	for i, tokens := range instances {
 		if len(tokens) == 0 {
 			continue
 		}
 
-		// We can safely append the input slice because elements inside are never shuffled.
-		h = append(h, tokens)
 		numTokens += len(tokens)
+		lists[i] = newTokenList(tokens...)
 	}
-	heap.Init(&h)
+
+	at := func(s *tokenList) uint32 { return s.At() }
+	less := func(a, b uint32) bool { return a < b }
+	close := func(s *tokenList) {}
+	tree := loser.New(lists, math.MaxUint32, at, less, close)
 
 	out := make([]uint32, 0, numTokens)
 
-	for h.Len() > 0 {
-		// The minimum element in the tree is the root, at index 0.
-		lowest := h[0]
-		out = append(out, lowest[0])
-
-		if len(lowest) > 1 {
-			// Remove the first token from the lowest because we popped it
-			// and then fix the heap to keep it sorted.
-			h[0] = h[0][1:]
-			heap.Fix(&h, 0)
-		} else {
-			heap.Remove(&h, 0)
-		}
+	for tree.Next() {
+		out = append(out, tree.Winner().At())
 	}
 
 	return out