--- a/ring/ring.go
+++ b/ring/ring.go
@@ -756,9 +756,7 @@ func (r *Ring) shuffleShard(identifier string, size int, lookbackPeriod time.Dur
 // mergeTokenGroups returns a sorted list of all tokens in each entry in groupsByName.
 // Each element of groupsByName is assumed to already be sorted.
 func mergeTokenGroups(groupsByName map[string][]uint32) []uint32 {
-	tokenCount := 0
-	groupsByIndex := make([][]uint32, 0, len(groupsByName))
-	nextIndex := make([]int, 0, len(groupsByName))
+	groupsList := make([][]uint32, 0, len(groupsByName))
 
 	for _, group := range groupsByName {
 		// If there's only one group, there's nothing to merge.
@@ -766,42 +764,10 @@ func mergeTokenGroups(groupsByName map[string][]uint32) []uint32 {
 			return group
 		}
 
-		tokenCount += len(group)
-		groupsByIndex = append(groupsByIndex, group)
-		nextIndex = append(nextIndex, 0)
+		groupsList = append(groupsList, group)
 	}
 
-	merged := make([]uint32, 0, tokenCount)
-
-	for i := 0; i < tokenCount; i++ {
-		haveSeenGroupWithRemainingToken := false
-		lowestToken := uint32(0)
-		lowestTokenGroupIndex := 0
-
-		for groupIndex, group := range groupsByIndex {
-			nextIndexInGroup := nextIndex[groupIndex]
-
-			if nextIndexInGroup >= len(group) {
-				continue
-			}
-
-			if group[nextIndexInGroup] < lowestToken || !haveSeenGroupWithRemainingToken {
-				lowestToken = group[nextIndexInGroup]
-				lowestTokenGroupIndex = groupIndex
-			}
-
-			haveSeenGroupWithRemainingToken = true
-		}
-
-		if !haveSeenGroupWithRemainingToken {
-			return merged
-		}
-
-		merged = append(merged, lowestToken)
-		nextIndex[lowestTokenGroupIndex]++
-	}
-
-	return merged
+	return MergeTokens(groupsList)
 }
 
 // GetInstanceState returns the current state of an instance or an error if the