… CHANGELOG.md
… go.sum
… ring/ring_test.go