@charleskorn 2023-04-12
> I'm assuming the loser tree implementation is correct, given it's getting copied from Loki 
I have modified it slightly to not use an interface, but the tests exercise the same test cases, so I believe this hasn't broken anything.

> I also assume the fuzz test has been run manually given it doesn't run in CI.
Yep.